pizza-api/pyproject.toml
[tool.poetry]
name = "pizza-api"
version = "0.1.0"
description = ""
authors = ["Himanshu <himanshchauhan6@gmail.com>"]
readme = "README.md"
packages = [{include = "pizza_api"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
fastapi-jwt-auth = "0.5.0"
werkzeug = "^3.0.1"
fastapi = "0.99.1"
pydantic = "1.10.11"
psycopg2-binary = "2.9.1"
sqlalchemy-utils = "0.37.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
